move receiver to file
