+++ src/core/core.cxx
@@ -25,3 +25,4 @@ export import :promise;
 export import :schedule;
 export import :root;
 export import :execute;
+export import :receiver;
